Move merge actions out of StorageDelegator
